// src/api/generated/schemas/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NervesHub API
 * OpenAPI spec version: 2.0.0
 */

/**
 * A registered user
 */
export interface User {
  /** Email address */
  email: string;
  /** User ID */
  id?: number;
  /** Creation timestamp */
  inserted_at?: string;
  /**
   * Users name
   * @pattern [a-zA-Z][a-zA-Z0-9_]+
   */
  name: string;
  /** Update timestamp */
  updated_at?: string;
}